+++ src/Misc/_api.ts
@@ -1,1 +1,9 @@
-export {Promisable} from './Promisable'
+/** @ignore *//** */
+
+import * as JSON from './JSON/_api'
+
+export {JSON}
+
+// LEGACY
+
+export {Promisable} from '../Any/Promisable'