+++ src/Object/AssignUp.ts
@@ -3,14 +3,14 @@ import {IterationOf} from '../Iteration/IterationOf'
 import {MergeUp} from './MergeUp'
 import {Pos} from '../Iteration/Pos'
 import {Next} from '../Iteration/Next'
-import {Length} from '../Tuple/Length'
+import {Length} from '../List/Length'
 import {Cast} from '../Any/Cast'
-import {Tuple} from '../Tuple/Tuple'
+import {List} from '../List/List'
 
 /**
  * @hidden
  */
-type _Assign<O extends object, Os extends Tuple<object>, I extends Iteration = IterationOf<'0'>> = {
+type _Assign<O extends object, Os extends List<object>, I extends Iteration = IterationOf<'0'>> = {
     0: _Assign<MergeUp<Os[Pos<I>], O>, Os, Next<I>>
     1: O
 }[
@@ -19,7 +19,7 @@ type _Assign<O extends object, Os extends Tuple<object>, I extends Iteration = I
     : 0
 ]
 
-/** Assign a list of **`object`** into **`O`** with **`MergeUp`** (last-in combines or overrides)
+/** Assign a list of **`object`** into **`O`** with [[MergeUp]] (last-in combines or overrides)
  * @param O to assign to
  * @param Os to assign
  * @returns **`object`**
@@ -27,7 +27,7 @@ type _Assign<O extends object, Os extends Tuple<object>, I extends Iteration = I
  * ```ts
  * ```
  */
-export type AssignUp<O extends object, Os extends Tuple<object>> =
+export type AssignUp<O extends object, Os extends List<object>> =
     _Assign<O, Os> extends infer X
     ? Cast<X, object>
     : never