+++ src/Object/Merge.ts
@@ -22,14 +22,14 @@ type MergeDeep<O, O1> = // we do not distribute this one => recursive distribute
 
 // If we wanted to dive in the tuples as well
 // Kind<(O | O1)> extends 'array'
-// ? TMerge<Cast<O, Tuple>, Cast<O1, Tuple>> extends infer M
+// ? TMerge<Cast<O, List>, Cast<O1, List>> extends infer M
 //   ? {[K in keyof M]: MergeDeep<M[K], At<O1 & {}, K>>}
 //   : never // this is a trick to force ts to do it deeply
 // : O       // in versions <= 3.8 no-recursive-conditional
 
 /** Complete the fields of **`O`** with the ones of **`O1`**
  * ('deep' option will skip nullable objects to be merged).
- * For more advanced capabilities, see **`MergeUp`**.
+ * For more advanced capabilities, see [[MergeUp]].
  * @param O to complete
  * @param O1 to copy from
  * @param depth to do it deeply (?=`'flat'`)