+++ src/Union/Has.ts
@@ -1,11 +1,11 @@
 import {Intersect} from './Intersect'
 import {True, False} from '../Boolean/Boolean'
 
-/** Check whether **`U`** contains **`A`** or not
- * (Note: **`U`** & **`A`** can be interchanged)
+/** Check whether **`U`** contains **`A`**
+ * (**`U`** & **`A`** can be interchanged)
  * @param U to be inspected
  * @param A to check within
- * @returns **`Boolean`**
+ * @returns [[Boolean]]
  * @example
  * ```ts
  * ```