+++ src/Union/IntersectOf.ts
@@ -1,6 +1,6 @@
 // Credit https://stackoverflow.com/a/50375286/3570903
 
-/** Transform a **union** to an **intersection**
+/** Transform a [[Union]] to an **intersection**
  * @param U to transform
  * @returns **intersection**
  * @example
@@ -8,6 +8,6 @@
  * ```
  */
 export type IntersectOf<U extends any> =
-    (U extends unknown ? (k: U) => void : never) extends ((k: infer I) => void)
+    (U extends unknown? (k: U) => void : never) extends ((k: infer I) => void)
     ? I
     : never