+++ src/Union/_api.ts
@@ -1,7 +1,10 @@
+/** @ignore *//** */
+
 export {Diff} from './Diff'
 export {Exclude} from './Exclude'
 export {Filter} from './Filter'
 export {Has} from './Has'
+export {HasAll} from './HasAll'
 export {Intersect} from './Intersect'
 export {IntersectOf} from './IntersectOf'
 export {Keys} from './Keys'
@@ -9,8 +12,14 @@ export {Last} from './Last'
 export {Merge} from './Merge'
 export {NonNullable} from './NonNullable'
 export {Nullable} from './Nullable'
+export {Omit} from './Omit'
+export {Pick} from './Pick'
 export {Pop} from './Pop'
 export {Replace} from './Replace'
 export {Select} from './Select'
 export {Strict} from './Strict'
-export {TupleOf} from './TupleOf'
+export {ListOf} from './ListOf'
+
+// LEGACY
+
+export {ListOf as TupleOf} from './ListOf'