+++ tst/Function.ts
@@ -7,13 +7,13 @@ const {checks, check} = Test
 //////////////////////////////////////////////////////////////////////////////////////////
 // FUNCTION //////////////////////////////////////////////////////////////////////////////
 
-const FN = (a: string, b: number, c: object) => true
+const fn = (a: string, b: number, c: object) => true
 
 // ---------------------------------------------------------------------------------------
 // ARROW
 
 checks([
-    check<F.Function<[string, number, object], boolean>,   typeof FN & Function,  Test.Pass>(),
+    check<F.Function<[string, number, object], boolean>,    typeof fn,  Test.Pass>(),
 ])
 
 // ---------------------------------------------------------------------------------------
@@ -21,24 +21,41 @@ checks([
 
 // sync ----------------------------------------------------------------------------------
 
-declare function compose<Fns extends F.Function[]>(...args: F.Composer<Fns>): F.Compose<Fns>
+declare function compozeSync<Fns extends F.Function[]>(...args: F.Composer<Fns>): F.Composed<Fns>
+declare const composeSync: F.Compose<'sync'>
 
-const composedSync = compose(
+const compozedSync = compozeSync(
     (message: string)                   => false,                   // receive previous return
     (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
     <T>(generic: T)                     => generic,                 // receive previous return
     (name: string, age: number)         => ({name, age}),           // receive parameters
 )
 
+const composedSync = composeSync(
+    (message: string)                   => false,                   // receive previous return
+    (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
+    <T>(generic: T)                     => generic,                 // receive previous return
+    (name: string, age: number)         => ({name, age}),            // receive parameters
+)
+
 checks([
-    check<(typeof composedSync),    (name: string, age: number) => boolean,  Test.Pass>(),
+    check<(typeof compozedSync),    (name: string, age: number) => boolean,     Test.Pass>(),
+    check<(typeof composedSync),    (name: string, age: number) => boolean,     Test.Pass>(),
 ])
 
 // async ---------------------------------------------------------------------------------
 
-declare function compose<Fns extends F.Function[]>(...args: F.Composer<Fns, 'async'>): F.Compose<Fns, 'async'>
+declare function compozeAsync<Fns extends F.Function[]>(...args: F.Composer<Fns, 'async'>): F.Composed<Fns, 'async'>
+declare const composeAsync: F.Compose<'async'>
 
-const composedAsync = compose(
+const compozedAsync = compozeAsync(
+    (message: string)                         => false,                   // receive previous return
+    async (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
+    async <T>(generic: T)                     => generic,                 // receive previous return
+    async (name: string, age: number)         => ({name, age}),           // receive parameters
+)
+
+const composedAsync = composeAsync(
     (message: string)                         => false,                   // receive previous return
     async (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
     async <T>(generic: T)                     => generic,                 // receive previous return
@@ -46,6 +63,7 @@ const composedAsync = compose(
 )
 
 checks([
+    check<(typeof compozedAsync),   (name: string, age: number) => Promise<boolean>,    Test.Pass>(),
     check<(typeof composedAsync),   (name: string, age: number) => Promise<boolean>,    Test.Pass>(),
 ])
 
@@ -64,18 +82,22 @@ const t = curried()
 const test00 = curried(__, 26)(__, true, 'JJ', __)('Jane', 'Jini') // boolean
 const test01 = curried('Jane', 26, true, __)('JJ', __)('Jini')     // boolean
 
+type t = typeof test01
+
+
+
 // ---------------------------------------------------------------------------------------
 // PARAMETERS
 
 checks([
-    check<F.Parameters<typeof FN>,    [string, number, object],   Test.Pass>(),
+    check<F.Parameters<typeof fn>,    [string, number, object],   Test.Pass>(),
 ])
 
 // ---------------------------------------------------------------------------------------
 // LENGTH
 
 checks([
-    check<F.Length<typeof FN>,                          3,          Test.Pass>(),
+    check<F.Length<typeof fn>,                          3,          Test.Pass>(),
     check<F.Length<(a1: any, a2?: any) => any>,         1 | 2,      Test.Pass>(),
     check<F.Length<(a1: any, a2?: any) => any, 's'>,    '1' | '2',  Test.Pass>(),
 ])
@@ -85,24 +107,41 @@ checks([
 
 // sync ----------------------------------------------------------------------------------
 
-declare function pipe<Fns extends F.Function[]>(...args: F.Piper<Fns>): F.Pipe<Fns>
+declare function pypeSync<Fns extends F.Function[]>(...args: F.Piper<Fns>): F.Piped<Fns>
+declare const pipeSync: F.Pipe<'sync'>
 
-const pipedSync = pipe(
+const pypedSync = pypeSync(
     (name: string, age: number)         => ({name, age}),           // receive parameters
     <T>(generic: T)                     => generic,                 // receive previous return
     (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
     (message: string)                   => false,                   // receive previous return
 )
 
+const pipedSync = pipeSync(
+    (name: string, age: number)           => ({name, age}),           // receive parameters
+    <T>(generic: T)                     => generic,                 // receive previous return
+    (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
+    (message: string)                   => false,                   // receive previous return
+)
+
 checks([
+    check<(typeof pypedSync),   (name: string, age: number) => boolean,  Test.Pass>(),
     check<(typeof pipedSync),   (name: string, age: number) => boolean,  Test.Pass>(),
 ])
 
 // async ---------------------------------------------------------------------------------
 
-declare function pipe<Fns extends F.Function[]>(...args: F.Piper<Fns, 'async'>): F.Pipe<Fns, 'async'>
+declare function pypeAsync<Fns extends F.Function[]>(...args: F.Piper<Fns, 'async'>): F.Piped<Fns, 'async'>
+declare const pipeAsync: F.Pipe<'async'>
 
-const pipedAsync = pipe(
+const pypedAsync = pypeAsync(
+    (name: string, age: number)               => ({name, age}),           // receive parameters
+    async <T>(generic: T)                     => generic,                 // receive previous return
+    async (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
+    (message: string)                         => false,                   // receive previous return
+)
+
+const pipedAsync = pipeAsync(
     (name: string, age: number)               => ({name, age}),           // receive parameters
     async <T>(generic: T)                     => generic,                 // receive previous return
     async (info: {name: string, age: number}) => `Welcome, ${info.name}`, // receive previous return
@@ -110,14 +149,23 @@ const pipedAsync = pipe(
 )
 
 checks([
+    check<(typeof pypedAsync),   (name: string, age: number) => Promise<boolean>,   Test.Pass>(),
     check<(typeof pipedAsync),   (name: string, age: number) => Promise<boolean>,   Test.Pass>(),
+])
+
+// ---------------------------------------------------------------------------------------
+// PROMISIFY
+
+checks([
+    check<F.Promisify<(typeof fn)>,                      (a: string, b: number, c: object) => Promise<boolean>,     Test.Pass>(),
+    check<F.Promisify<(a: string) => Promise<boolean>>,  (a: string) => Promise<boolean>,                           Test.Pass>(),
 ])
 
 // ---------------------------------------------------------------------------------------
 // RETURN
 
 checks([
-    check<F.Return<typeof FN>,    boolean,    Test.Pass>(),
+    check<F.Return<typeof fn>,    boolean,    Test.Pass>(),
 ])
 
 // ---------------------------------------------------------------------------------------